wip: fixing tutorial
